♻️ refactor: move bvh_internal.h from public to source directory
bvh_internal.h is only used by 4 files within src/spatial/bvh/
but was exposed in includes/spatial/ (public headers). Move it
to src/spatial/bvh/ alongside its consumers, matching the pattern
already used by bvhd_internal.h in src/spatial/debug/.
