+++ src/spatial/bvh/bvh_internal.h
